refactor: drop the Learner and Teacher semantic from User
