+++ .idea/jpb-settings.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="DatabaseMigrationSettings" lastSelectedDirectory="$PROJECT_DIR$/server/src/main/db.migrations">
+  <component name="DatabaseMigrationSettings" lastSelectedDirectory="$PROJECT_DIR$/server/src/main/resources/db/migration">
     <database-infos>
       <database-info>
         <dbms-id>postgres</dbms-id>
