--- server/src/main/kotlin/org/elaastix/server/users/cohorts/entities/AbstractCohortEntity.kt
@@ -1,55 +1,0 @@
-/*
- * Elaastic / ElaastiX - formative assessment system
- * Copyright (C) 2019  Université de Toulouse and Université Toulouse Capitole.
- * SPDX-License-Identifier: AGPL-3.0-or-later
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package org.elaastix.server.users.cohorts.entities
-
-import jakarta.persistence.DiscriminatorColumn
-import jakarta.persistence.Entity
-import jakarta.persistence.Inheritance
-import jakarta.persistence.InheritanceType
-import jakarta.persistence.ManyToMany
-import jakarta.persistence.Table
-import jakarta.validation.constraints.NotBlank
-import jakarta.validation.constraints.NotNull
-import jakarta.validation.constraints.Size
-import org.elaastix.commons.jpa.AbstractEntity
-import org.elaastix.mm.users.User
-import org.elaastix.mm.users.cohorts.Cohort
-import org.elaastix.server.users.entities.AbstractUserEntity
-
-/**
- * Abstract entity for cohorts.
- * Defines the JPA inheritance strategy used for specialised entities.
- *
- * @see [Cohort]
- */
-@Entity
-@Table(name = "cohorts")
-@Inheritance(strategy = InheritanceType.SINGLE_TABLE)
-@DiscriminatorColumn(name = "cohortType", length = 16)
-abstract class AbstractCohortEntity<T : User>(
-    @NotNull
-    @NotBlank
-    @Size(min = 2, max = 64)
-    override var name: String,
-
-    @ManyToMany
-    override var administrators: MutableSet<AbstractUserEntity>,
-) : AbstractEntity(),
-    Cohort<T>
