--- server/src/main/kotlin/org/elaastix/server/users/cohorts/entities/TeacherCohortEntity.kt
@@ -1,45 +1,0 @@
-/*
- * Elaastic / ElaastiX - formative assessment system
- * Copyright (C) 2019  Université de Toulouse and Université Toulouse Capitole.
- * SPDX-License-Identifier: AGPL-3.0-or-later
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package org.elaastix.server.users.cohorts.entities
-
-import jakarta.persistence.DiscriminatorValue
-import jakarta.persistence.Entity
-import jakarta.persistence.ManyToMany
-import org.elaastix.mm.users.Teacher
-import org.elaastix.mm.users.cohorts.TeacherCohort
-import org.elaastix.server.users.entities.AbstractUserEntity
-import org.elaastix.server.users.entities.TeacherEntity
-
-/**
- * @see [TeacherCohort]
- */
-@Entity
-@DiscriminatorValue("TEACHER")
-class TeacherCohortEntity(
-    name: String,
-    administrators: MutableSet<AbstractUserEntity>,
-
-    @ManyToMany
-    override var members: MutableSet<TeacherEntity>,
-) : AbstractCohortEntity<Teacher>(
-        name,
-        administrators,
-    ),
-    TeacherCohort
