+++ server/src/main/resources/db/migration/V1__Initial.sql
@@ -19,45 +19,45 @@
 
 CREATE TABLE cohorts
 (
-    id          UUID                        NOT NULL,
-    updated_at  TIMESTAMP WITHOUT TIME ZONE NOT NULL,
-    version     BIGINT                      NOT NULL,
-    cohort_type VARCHAR(255)                NOT NULL,
-    name        VARCHAR(64)                 NOT NULL,
+    id         UUID                        NOT NULL,
+    updated_at TIMESTAMP WITHOUT TIME ZONE NOT NULL,
+    version    BIGINT                      NOT NULL,
+    name       VARCHAR(64)                 NOT NULL,
     CONSTRAINT pk_cohorts PRIMARY KEY (id)
 );
 
 CREATE TABLE cohorts_administrators
 (
-    abstract_cohort_entity_id UUID NOT NULL,
-    administrators_id         UUID NOT NULL,
-    CONSTRAINT pk_cohorts_administrators PRIMARY KEY (abstract_cohort_entity_id, administrators_id)
+    cohort_entity_id  UUID NOT NULL,
+    administrators_id UUID NOT NULL,
+    CONSTRAINT pk_cohorts_administrators PRIMARY KEY (cohort_entity_id, administrators_id)
 );
 
 CREATE TABLE cohorts_members
 (
-    teacher_cohort_entity_id UUID NOT NULL,
-    members_id               UUID NOT NULL,
-    CONSTRAINT pk_cohorts_members PRIMARY KEY (teacher_cohort_entity_id, members_id)
+    cohort_entity_id UUID NOT NULL,
+    members_id       UUID NOT NULL,
+    CONSTRAINT pk_cohorts_members PRIMARY KEY (cohort_entity_id, members_id)
 );
 
 CREATE TABLE users
 (
-    id         UUID                        NOT NULL,
-    updated_at TIMESTAMP WITHOUT TIME ZONE NOT NULL,
-    version    BIGINT                      NOT NULL,
-    user_type  VARCHAR(255)                NOT NULL,
+    id                     UUID                        NOT NULL,
+    updated_at             TIMESTAMP WITHOUT TIME ZONE NOT NULL,
+    version                BIGINT                      NOT NULL,
+    is_writer_mode_enabled BOOLEAN                     NOT NULL,
+    is_administrator       BOOLEAN                     NOT NULL,
     CONSTRAINT pk_users PRIMARY KEY (id)
 );
 
 ALTER TABLE cohorts_administrators
-    ADD CONSTRAINT fk_cohadm_on_abstract_cohort_entity FOREIGN KEY (abstract_cohort_entity_id) REFERENCES cohorts (id);
+    ADD CONSTRAINT fk_cohadm_on_cohort_entity FOREIGN KEY (cohort_entity_id) REFERENCES cohorts (id);
 
 ALTER TABLE cohorts_administrators
-    ADD CONSTRAINT fk_cohadm_on_abstract_user_entity FOREIGN KEY (administrators_id) REFERENCES users (id);
+    ADD CONSTRAINT fk_cohadm_on_user_entity FOREIGN KEY (administrators_id) REFERENCES users (id);
 
 ALTER TABLE cohorts_members
-    ADD CONSTRAINT fk_cohmem_on_teacher_cohort_entity FOREIGN KEY (teacher_cohort_entity_id) REFERENCES cohorts (id);
+    ADD CONSTRAINT fk_cohmem_on_cohort_entity FOREIGN KEY (cohort_entity_id) REFERENCES cohorts (id);
 
 ALTER TABLE cohorts_members
-    ADD CONSTRAINT fk_cohmem_on_teacher_entity FOREIGN KEY (members_id) REFERENCES users (id);
+    ADD CONSTRAINT fk_cohmem_on_user_entity FOREIGN KEY (members_id) REFERENCES users (id);
